+++ src/constants/regex.ts
@@ -1,5 +1,6 @@
 export const URL_REGEX =
     /(https:\/\/|http:\/\/)?(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)/gu;
+
 export const MENTION_REGEX = /@[^\s]+/g;
 
 export const HASHTAG_REGEX = /(#\w*[A-Za-z]\w*)/g;
